feature/MTCE-226-Monorepo-v6-refactoring-EasyEventDispatcher
