+++ composer.json
@@ -108,6 +108,8 @@
             "EonX\\EasyEncryption\\": "packages/EasyEncryption/src",
             "EonX\\EasyErrorHandler\\": "packages/EasyErrorHandler/src",
             "EonX\\EasyEventDispatcher\\": "packages/EasyEventDispatcher/src",
+            "EonX\\EasyEventDispatcher\\Bundle\\": "packages/EasyEventDispatcher/bundle",
+            "EonX\\EasyEventDispatcher\\Laravel\\": "packages/EasyEventDispatcher/laravel",
             "EonX\\EasyHttpClient\\": "packages/EasyHttpClient/src",
             "EonX\\EasyLock\\": "packages/EasyLock/src",
             "EonX\\EasyLogging\\": "packages/EasyLogging/src",
@@ -154,6 +156,8 @@
             "EonX\\EasyErrorHandler\\Tests\\": "packages/EasyErrorHandler/tests",
             "EonX\\EasyErrorHandler\\Tests\\Bridge\\Symfony\\Fixtures\\App\\": "packages/EasyErrorHandler/tests/Bridge/Symfony/Fixtures/app/src",
             "EonX\\EasyEventDispatcher\\Tests\\": "packages/EasyEventDispatcher/tests",
+            "EonX\\EasyEventDispatcher\\Tests\\Unit\\Bundle\\": "packages/EasyEventDispatcher/tests/Unit/bundle",
+            "EonX\\EasyEventDispatcher\\Tests\\Unit\\Laravel\\": "packages/EasyEventDispatcher/tests/Unit/laravel",
             "EonX\\EasyHttpClient\\Tests\\": "packages/EasyHttpClient/tests",
             "EonX\\EasyHttpClient\\Tests\\Bridge\\Symfony\\Fixtures\\App\\": "packages/EasyHttpClient/tests/Bridge/Symfony/Fixtures/app/src",
             "EonX\\EasyLock\\Tests\\": "packages/EasyLock/tests",
