+++ packages/EasyActivity/tests/Fixture/app/config/bundles.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use EonX\EasyActivity\Bundle\EasyActivityBundle;
 use EonX\EasyDoctrine\Bridge\Symfony\EasyDoctrineSymfonyBundle;
-use EonX\EasyEventDispatcher\Bridge\Symfony\EasyEventDispatcherSymfonyBundle;
+use EonX\EasyEventDispatcher\Bundle\EasyEventDispatcherBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 
 return [
     EasyActivityBundle::class => [
         'all' => true,
     ],
-    EasyEventDispatcherSymfonyBundle::class => [
+    EasyEventDispatcherBundle::class => [
         'all' => true,
     ],
     EasyDoctrineSymfonyBundle::class => [
