+++ packages/EasyAsync/tests/Stub/EventDispatcherStub.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyAsync\Tests\Stub;
 
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
 
 final class EventDispatcherStub implements EventDispatcherInterface
 {
