+++ packages/EasyAsync/tests/Stub/KernelStub.php
@@ -6,7 +6,7 @@ namespace EonX\EasyAsync\Tests\Stub;
 use Doctrine\DBAL\Connection;
 use Doctrine\Persistence\ManagerRegistry;
 use EonX\EasyAsync\Bundle\EasyAsyncBundle;
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
 use EonX\EasyRandom\Bridge\Symfony\EasyRandomSymfonyBundle;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
