+++ packages/EasyEventDispatcher/bundle/EasyEventDispatcherBundle.php
@@ -1,16 +1,14 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyEventDispatcher\Bridge\Symfony;
+namespace EonX\EasyEventDispatcher\Bundle;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\HttpKernel\Bundle\AbstractBundle;
 
-final class EasyEventDispatcherSymfonyBundle extends AbstractBundle
+final class EasyEventDispatcherBundle extends AbstractBundle
 {
-    protected string $extensionAlias = 'easy_event_dispatcher';
-
     public function __construct()
     {
         $this->path = \realpath(__DIR__);
@@ -18,6 +16,6 @@ final class EasyEventDispatcherSymfonyBundle extends AbstractBundle
 
     public function loadExtension(array $config, ContainerConfigurator $container, ContainerBuilder $builder): void
     {
-        $container->import(__DIR__ . '/Resources/config/services.php');
+        $container->import('config/services.php');
     }
 }
