+++ packages/EasyEventDispatcher/bundle/config/services.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use EonX\EasyEventDispatcher\Bridge\Symfony\EventDispatcher;
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcher;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface as SymfonyEventDispatcherInterface;
 
 return static function (ContainerConfigurator $container): void {
