+++ packages/EasyEventDispatcher/docs/laravel/install.md
@@ -16,7 +16,7 @@ not familiar with this concept make sure to have a look at the [documentation][1
 'providers' => [
     // Other Service Providers...
 
-    \EonX\EasyEventDispatcher\Bridge\Laravel\EasyEventDispatcherServiceProvider::class,
+    \EonX\EasyEventDispatcher\Laravel\EasyEventDispatcherServiceProvider::class,
 ],
 ```
 
@@ -35,7 +35,7 @@ $app = new Laravel\Lumen\Application(\dirname(__DIR__));
 
 // Other actions...
 
-$app->register(\EonX\EasyEventDispatcher\Bridge\Laravel\EasyEventDispatcherServiceProvider::class);
+$app->register(\EonX\EasyEventDispatcher\Laravel\EasyEventDispatcherServiceProvider::class);
 ```
 
 [1]: https://laravel.com/docs/5.8/providers
