+++ packages/EasyEventDispatcher/tests/Unit/AbstractUnitTestCase.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyEventDispatcher\Tests;
+namespace EonX\EasyEventDispatcher\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
@@ -10,12 +10,12 @@ use Symfony\Component\Filesystem\Filesystem;
  * This class has for objective to provide common features to all tests without having to update
  * the class they all extend.
  */
-abstract class AbstractTestCase extends TestCase
+abstract class AbstractUnitTestCase extends TestCase
 {
     protected function tearDown(): void
     {
         $fs = new Filesystem();
-        $var = __DIR__ . '/../var';
+        $var = __DIR__ . '/../../var';
 
         if ($fs->exists($var)) {
             $fs->remove($var);
