+++ packages/EasyEventDispatcher/tests/Unit/laravel/EasyEventDispatcherServiceProviderTest.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyEventDispatcher\Tests\Bridge\Laravel;
+namespace EonX\EasyEventDispatcher\Tests\Unit\Laravel;
 
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
-use EonX\EasyEventDispatcher\Tests\Bridge\Symfony\Stubs\EventStub;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Tests\Stub\Even\EventStub;
 use Illuminate\Contracts\Events\Dispatcher as IlluminateDispatcherContract;
 
 final class EasyEventDispatcherServiceProviderTest extends AbstractLaravelTestCase
@@ -12,7 +12,7 @@ final class EasyEventDispatcherServiceProviderTest extends AbstractLaravelTestCa
     public function testDispatchInLaravel(): void
     {
         $app = $this->getApp();
-        /** @var \EonX\EasyEventDispatcher\Tests\Bridge\Laravel\Stubs\LaravelEventDispatcherStub $symfonyDispatcher */
+        /** @var \EonX\EasyEventDispatcher\Tests\Stub\Dispatcher\LaravelEventDispatcherStub $symfonyDispatcher */
         $symfonyDispatcher = $app->get(IlluminateDispatcherContract::class);
         $easyDispatcher = $app->get(EventDispatcherInterface::class);
         $event = new EventStub();
