+++ packages/EasyHttpClient/tests/Bridge/Laravel/EasyHttpClientServiceProviderTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyHttpClient\Tests\Bridge\Laravel;
 
-use EonX\EasyEventDispatcher\Bridge\Laravel\EasyEventDispatcherServiceProvider;
+use EonX\EasyEventDispatcher\Laravel\EasyEventDispatcherServiceProvider;
 use EonX\EasyHttpClient\Implementations\Symfony\WithEventsHttpClient;
 use EonX\EasyWebhook\Bridge\BridgeConstantsInterface as EasyWebhookBridgeConstantsInterface;
 use EonX\EasyWebhook\Bridge\Laravel\EasyWebhookServiceProvider;
