+++ packages/EasyHttpClient/tests/Bridge/Symfony/Fixtures/app/config/services.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
 use EonX\EasyHttpClient\Tests\Bridge\Symfony\Fixtures\App\Client\SomeClient;
 use EonX\EasyHttpClient\Tests\Stubs\EventDispatcherStub;
 use EonX\EasyTest\HttpClient\TestResponseFactory;
