+++ packages/EasyWebhook/tests/Bridge/Laravel/AbstractLumenTestCase.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyWebhook\Tests\Bridge\Laravel;
 
-use EonX\EasyEventDispatcher\Bridge\Laravel\EasyEventDispatcherServiceProvider;
+use EonX\EasyEventDispatcher\Laravel\EasyEventDispatcherServiceProvider;
 use EonX\EasyLock\Bridge\Laravel\EasyLockServiceProvider;
 use EonX\EasyWebhook\Bridge\Laravel\EasyWebhookServiceProvider;
 use EonX\EasyWebhook\Tests\AbstractTestCase;
