+++ packages/EasyWebhook/tests/Stubs/EventDispatcherStub.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyWebhook\Tests\Stubs;
 
-use EonX\EasyEventDispatcher\Interfaces\EventDispatcherInterface;
+use EonX\EasyEventDispatcher\Dispatcher\EventDispatcherInterface;
 use EonX\EasyWebhook\Interfaces\WebhookEventInterface;
 use InvalidArgumentException;
